Fix default resampling quality selection
When the user doesn't explicitly specify the resampling quality using the `-Q`
flag, the quality should be calculated based on Jack's server sampling rate.

This isn't the case, because the `rqual` variable is always set to 16 in
`jack_initialize`. This results in the quality always being 16.

To fix this and make the behavior the same as the `zita-ajbridge` tools, the
redundant `rqual` clamping of the value to the range `[16,96]` should be
removed from `jack_initialize()`, because this is done later in
`jack_initialize_part2()` only AFTER the quality calculation.